+++ enterprise/__init__.py
@@ -2,6 +2,6 @@
 Your project description goes here.
 """
 
-__version__ = "3.61.14"
+__version__ = "4.0.0"
 
 default_app_config = "enterprise.apps.EnterpriseConfig"